% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics_plot.bigtime.VARMA}
\alias{diagnostics_plot.bigtime.VARMA}
\title{diagnostics_plot function for VARMA models}
\usage{
\method{diagnostics_plot}{bigtime.VARMA}(mod, variable = 1, dates = NULL)
}
\arguments{
\item{mod}{VAR model estimated using sparseVARMA}

\item{variable}{Variable to show. Either numeric (which column) or character
(variable name)}

\item{dates}{Optional Date vector.}
}
\description{
diagnostics_plot function for VARMA models
}
