% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{.diagnostics_plot}
\alias{.diagnostics_plot}
\title{Internal function to plot the diagnostics plot}
\usage{
.diagnostics_plot(Y, fit, res, s, variable, dates)
}
\arguments{
\item{Y}{observed values}

\item{fit}{fitted values}

\item{res}{residual values}

\item{s}{how many observations were lost}

\item{variable}{variable to plot: either numeric or colname}

\item{dates}{vector of dates of length nrow(Y)}
}
\description{
Internal function to plot the diagnostics plot
}
\keyword{internal}
