% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_criteria.R
\name{.get_ic_vals}
\alias{.get_ic_vals}
\title{Calculates ICs}
\usage{
.get_ic_vals(res, tt, df)
}
\arguments{
\item{res}{Matrix of residuals}

\item{tt}{number of time periods}

\item{df}{degrees of freedom}
}
\description{
Not meant to be called by the user
}
\keyword{internal}
