% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexample.R
\docType{data}
\name{example}
\alias{example}
\title{Multivariate Time Series Example}
\format{A matrix with 100 rows and 15 columns}
\usage{
example
}
\description{
A matrix containing 15 time series in its columns that are each observed over 100 time points
}
\keyword{datasets}
