% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{fitted.bigtime.VARMA}
\alias{fitted.bigtime.VARMA}
\title{Gives the fitted values of a model estimated using
\code{\link{sparseVARMA}}}
\usage{
\method{fitted}{bigtime.VARMA}(object, ...)
}
\arguments{
\item{object}{Model estimated using \code{\link{sparseVARMA}}}

\item{...}{Not currently used}
}
\value{
Returns a matrix of fitted values
data(varma.example)
varma <- sparseVARMA(Y = scale(Y.varma), VARMAselection="cv")
f <- fitted(varma)
}
\description{
Gives the fitted values of a model estimated using
\code{\link{sparseVARMA}}
}
