% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_criteria.R
\name{get_ic_vals.bigtime.VAR}
\alias{get_ic_vals.bigtime.VAR}
\title{Calculates the Information Criteria for a model estimated using sparseVAR}
\usage{
\method{get_ic_vals}{bigtime.VAR}(mod, verbose = TRUE)
}
\arguments{
\item{mod}{Model estimated using sparseVAR}

\item{verbose}{Should information about the optimal selection be printed?}
}
\value{
Returns a list containing
\item{ics}{Values of the ICs for all lambdas}
\item{mins}{Which ic lead to the minimum (the row number)}
\item{selected_lambdas}{Which lambdas were selected}
}
\description{
The number of non-zero coefficients are taken as the degrees of freedom.
This is not valid in cases in which p>N
}
\examples{
dat <- simVAR(200, 2, 5, decay = 0.01)
mod <- sparseVAR(scale(dat$Y))
ics <- get_ic_vals(mod)
}
