% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_criteria.R
\name{get_ic_vals}
\alias{get_ic_vals}
\title{Calculates the Information Criteria for a VAR, VARX, VARMA model}
\usage{
get_ic_vals(mod, verbose = TRUE)
}
\arguments{
\item{mod}{Model estimated using sparseVAR, sparseVARX, or sparseVARMA}

\item{verbose}{Should information about the optimal selection be printed?}
}
\description{
The number of non-zero coefficients are taken as the degrees of freedom.
This is not valid in cases in which p>N
}
\examples{
dat <- simVAR(200, 2, 5, decay = 0.01)
mod <- sparseVAR(scale(dat$Y))
ics <- get_ic_vals(mod)
}
