% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_criteria.R
\name{ic_selection}
\alias{ic_selection}
\title{Selects the optimal penalty parameter using information criteria}
\usage{
ic_selection(mod, ic = c("bic", "aic", "hq"), verbose = FALSE)
}
\arguments{
\item{mod}{Model estimated using \code{sparseVAR}, \code{sparseVARX},
or \code{sparseVARMA}}

\item{ic}{Which information criteria should be used. Must be one of
\code{"bic"}, \code{"aic"} or \code{"hq"}}

\item{verbose}{If true, some useful information will be printed during the process}
}
\value{
Returns a model that uses the optimal penalty
}
\description{
Selects the optimal penalty parameter using information criteria
}
