% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simVAR.R
\name{plot.bigtime.simVAR}
\alias{plot.bigtime.simVAR}
\title{Plots a simulated VAR}
\usage{
\method{plot}{bigtime.simVAR}(x, ...)
}
\arguments{
\item{x}{Simulated data of class \code{bigtime.simVAR} obtained
from the \code{\link{simVAR}} function}

\item{...}{Not currently used}
}
\value{
Returns a ggplot2 plot
}
\description{
Plots a simulated VAR
}
