% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cv.R
\name{plot_cv}
\alias{plot_cv}
\title{Plot the Cross Validation Error Curve for a Sparse VAR or VARX}
\usage{
plot_cv(fit, ...)
}
\arguments{
\item{fit}{Fitted VAR, VARMA or VARX model.
returned by \code{\link{sparseVAR}}, \code{\link{sparseVARMA}} or \code{\link{sparseVARX}}.}

\item{...}{Not currently used}
}
\description{
Plot the Cross Validation Error Curve for a Sparse VAR or VARX
}
