% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{resid.bigtime.VAR}
\alias{resid.bigtime.VAR}
\title{Gives the residuals for VAR models estimated using sparseVAR}
\usage{
\method{resid}{bigtime.VAR}(object, ...)
}
\arguments{
\item{object}{model estimated using sparseVAR}

\item{...}{Not currently used}
}
\value{
Returns a matrix of residuals.
}
\description{
Gives the residuals for VAR models estimated using sparseVAR
}
