% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{residuals.bigtime.VARMA}
\alias{residuals.bigtime.VARMA}
\title{Gives the residuals for VARMA models estimated using sparseVARMA}
\usage{
\method{residuals}{bigtime.VARMA}(object, ...)
}
\arguments{
\item{object}{model estimated using sparseVARMA}

\item{...}{Not currently used}
}
\value{
Returns a matrix of residuals.
}
\description{
Gives the residuals for VARMA models estimated using sparseVARMA
}
