% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{residuals.bigtime.VARX}
\alias{residuals.bigtime.VARX}
\title{Gives the residuals for VARX models estimated using
\code{\link{sparseVARX}}}
\usage{
\method{residuals}{bigtime.VARX}(object, ...)
}
\arguments{
\item{object}{Model estimated using \code{\link{sparseVARX}}}

\item{...}{Not currently used}
}
\value{
Returns a matrix of residuals.
}
\description{
Gives the residuals for VARX models estimated using
\code{\link{sparseVARX}}
}
\examples{
data(varx.example)
varx <- sparseVARX(Y=scale(Y.varx), X=scale(X.varx), selection="cv")
res <- residuals(varx)
}
