% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simVAR.R
\name{simVAR}
\alias{simVAR}
\title{Simulates a VAR(p) with various sparsity patterns}
\usage{
simVAR(
  periods,
  k,
  p,
  coef_mat = NULL,
  const = rep(0, k),
  e_dist = rnorm,
  init_y = rep(0, k * p),
  max_abs_eigval = 0.8,
  burnin = periods,
  sparsity_pattern = c("none", "lasso", "hvar"),
  sparsity_options = NULL,
  decay = 1/p,
  seed = NULL,
  ...
)
}
\arguments{
\item{periods}{Number of periods to simulate}

\item{k}{Number of time series}

\item{p}{Maximum lag number. In case of sparsity_patter="none" this will be
the actual number of lags for all variables}

\item{coef_mat}{Coefficient matrix in companion form. If not provided,
one will be simulated}

\item{const}{constent term of VAR. Default is zero}

\item{e_dist}{Either a function taking argument n indicating the number of
variables in the system, or a matrix of dimensions k x (periods+burnin)}

\item{init_y}{Initial values. Defaults to zero}

\item{max_abs_eigval}{Maximum Eigenvalue of companion matrix. Only applicable
if coefficient matrix is being simulated}

\item{burnin}{Number of periods to be used for burnin}

\item{sparsity_pattern}{The sparsity pattern that should be simulated.
Options are: none for a dense VAR, lasso for a VAR with random zeroes,
and HVAR for an elementwise hirichical sparsity pattern}

\item{sparsity_options}{Named list of additional options for
when sparsity pattern is lasso or hvar. For lasso the option num_zero
determines the number of zeros. For hvar, the options zero_min (zero_max)
give the minimum (maximum) of zeroes for each variable in each equation,
and the option zeroes_in_self (boolean) determines if any of the
cofficients of a variable on itself should be zero.}

\item{decay}{How much smaller should parameters for laters lags be. The
smaller, the larger will early parameters be w.r.t. later ones.}

\item{seed}{Seed to be used for the simulation}

\item{...}{additional arguments passed to e_dist}
}
\value{
Returns an object of S3 class bigtime.simVAR containing the following
\item{Y}{Simulated Data}
\item{periods}{Number of periods of simulation}
\item{k}{Number of endogenous variables}
\item{p}{Maximum lag length; might be shorter due to sparsity patterns}
\item{coef_mat}{Coefficient Matrix used}
\item{is_coef_mat_simulated}{TRUE if the coef_mat was simulted, FALSE if
it was user provided}
\item{const}{Constant term}
\item{e_dist}{Errors used in the construction of the data}
\item{init_y}{Initial conditions}
\item{max_abs_eigval}{Maximum eigenvalue to which the companion matrix
was constraint}
\item{burnin}{Burnin period used}
\item{sparsity_pattern}{Which sparsity pattern was used}
\item{sparsity_options}{Extra options for the sparsity patterns that were used}
\item{seed}{seed used for the simulation}
}
\description{
Simulates a VAR(p) with various sparsity patterns
}
