% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simVAR.R
\name{summary.bigtime.simVAR}
\alias{summary.bigtime.simVAR}
\title{Gives a small summary of a VAR simulation}
\usage{
\method{summary}{bigtime.simVAR}(object, plot = TRUE, ...)
}
\arguments{
\item{object}{Simulated data of class \code{bigtime.simVAR} obtained
from the \code{\link{simVAR}} function}

\item{plot}{Should the VAR be plotted. Default is \code{TRUE}}

\item{...}{Not currently used}
}
\value{
If `plot=TRUE`, then a ggplot2 plot will be returned
}
\description{
Gives a small summary of a VAR simulation
}
