% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexample.R
\docType{data}
\name{X}
\alias{X}
\title{Multivariate Time Series Example}
\format{
A matrix with 50 rows and 3 columns
}
\usage{
X
}
\description{
A matrix containing three time series in its columns that are each observed over 50 time points
}
\keyword{datasets}
