% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexample.R
\docType{data}
\name{X.varx}
\alias{X.varx}
\title{VARX Time Series Example (\code{varx.example})}
\format{
Two matrices, \code{X.varx} and \code{Y.varx}, both of dimension 200x3
}
\usage{
X.varx
}
\description{
The data consists of a 200x3 matrix of endogenous variables, \code{Y.varx},
and a 200x3 matrix of exogenous variables, \code{X.varx}.
}
\keyword{datasets}
