% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexample.R
\docType{data}
\name{Y}
\alias{Y}
\title{Multivariate Time Series Example}
\format{A matrix with 50 rows and 2 columns}
\usage{
Y
}
\description{
A matrix containing two time series in its columns that are each observed over 50 time points
}
\keyword{datasets}
