% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexample.R
\docType{data}
\name{Y.var}
\alias{Y.var}
\title{VAR Time Series Example (\code{var.example})}
\format{
A matrix of dimension 200x5
}
\usage{
Y.var
}
\description{
The data consists of a 200x5 data matrix, \code{Y.var}, and was simulated from a sparse
VAR model with HLag sparsity pattern.
}
\keyword{datasets}
