% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexample.R
\docType{data}
\name{Y.varma}
\alias{Y.varma}
\title{VARMA Time Series Example (\code{varma.example})}
\format{
A matrix of dimension 200x3
}
\usage{
Y.varma
}
\description{
The data consists of a 200x3 data matrix, \code{Y.varma}, and was simulated from a
sparse VARMA model.
}
\keyword{datasets}
