% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simVAR.R
\name{create_rand_coef_mat}
\alias{create_rand_coef_mat}
\title{Creates a random coefficient matrix}
\usage{
create_rand_coef_mat(
  k,
  p,
  dist = runif,
  max_abs_eigval = 0.8,
  sparsity_pattern = c("none", "lasso", "hvar"),
  sparsity_options = NULL,
  decay = 0.5,
  ...
)
}
\arguments{
\item{k}{number of time series}

\item{p}{number of lags}

\item{dist}{distribution to draw coefficients from; must take n as argument
indicating number of draws wanted and must return one value per draw wanted,
e.g. no vector/matrix returns. Default is uniform distribution: Not currently used}

\item{max_abs_eigval}{if < 1, then var will be stable}

\item{sparsity_pattern}{The sparsity pattern that should be simulated.
Options are: none for a dense VAR, lasso for a VAR with random zeroes,
and HVAR for an elementwise hirichical sparsity pattern}

\item{sparsity_options}{Named list of additional options for
when sparsity pattern is lasso or hvar. For lasso the option num_zero
determines the number of zeros. For hvar, the options zero_min (zero_max)
give the minimum (maximum) of zeroes for each variable in each equation,
and the option zeroes_in_self (boolean) determines if any of the
cofficients of a variable on itself should be zero.}

\item{decay}{How fast should coefficient shrink when the lag increases.}

\item{...}{additional arguments forwarded to dist}
}
\value{
Returns a coefficient matrix in companion form.
}
\description{
Creates a random coefficient matrix
}
