% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics_plot.bigtime.VARX}
\alias{diagnostics_plot.bigtime.VARX}
\title{diagnostics_plot function for VARX models}
\usage{
\method{diagnostics_plot}{bigtime.VARX}(mod, variable = 1, dates = NULL)
}
\arguments{
\item{mod}{VARX model estimated using \code{\link{sparseVARX}}}

\item{variable}{Variable to show. Either numeric (which column) or character
(variable name)}

\item{dates}{Optional Date vector.}
}
\description{
Not supposed to be called directly. Rather call \code{\link{diagnostics_plot}}
}
