% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics_plot}
\alias{diagnostics_plot}
\title{Creates a diagnostic Plot}
\usage{
diagnostics_plot(mod, variable = 1, dates = NULL)
}
\arguments{
\item{mod}{VAR model estimated using sparseVAR}

\item{variable}{Variable to show. Either numeric (which column) or character
(variable name)}

\item{dates}{Optional Date vector.}
}
\value{
Returns a ggplot2 plot
}
\description{
Creates a diagnostic Plot
}
\examples{
dat <- simVAR(200, 2, 5, decay = 0.1, seed = 6150533,
                       sparsity_pattern = "hvar")
mod <- sparseVAR(dat$Y, h = 1)
diagnostics_plot(mod, variable = 1)
}
