% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directforecast.R
\name{directforecast}
\alias{directforecast}
\title{Function to obtain h-step ahead direct forecast based on estimated VAR, VARX or VARMA model}
\usage{
directforecast(fit, model, h = 1)
}
\arguments{
\item{fit}{Fitted sparse var, varx or varma model.}

\item{model}{Type of model that was estimated: var, varx or varma.}

\item{h}{Desired forecast horizon.}
}
\value{
A list with the following components
\item{Yhat}{h-step ahead forecasts for the k time series}
}
\description{
Function to obtain h-step ahead direct forecast based on estimated VAR, VARX or VARMA model
}
\examples{
data(Y)
varfit <- sparsevar(Y) # sparse VAR
forecasts <- directforecast(fit=varfit, model="var", h=1)
}
