% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{fitted.bigtime.VAR}
\alias{fitted.bigtime.VAR}
\title{Gives the fitted values of a model estimated using sparseVAR}
\usage{
\method{fitted}{bigtime.VAR}(object, ...)
}
\arguments{
\item{object}{model estimated using sparseVAR}

\item{...}{Not currently used}
}
\value{
Returns a matrix of fitted values
}
\description{
Gives the fitted values of a model estimated using sparseVAR
}
\examples{
dat <- bigtime::simVAR(200, 2, 5, decay = 0.001, seed = 6150533)
mod <- bigtime::sparseVAR(dat$Y)
f <- fitted(mod)
res <- resid(mod)

}
