% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{fitted.bigtime.VAR}
\alias{fitted.bigtime.VAR}
\title{Gives the fitted values of a model estimated using
\code{\link{sparseVAR}}}
\usage{
\method{fitted}{bigtime.VAR}(object, ...)
}
\arguments{
\item{object}{Model estimated using \code{\link{sparseVAR}}}

\item{...}{Not currently used}
}
\value{
Returns a matrix of fitted values
}
\description{
Gives the fitted values of a model estimated using
\code{\link{sparseVAR}}
}
\examples{
dat <- simVAR(periods=200, k=2, p=5, decay = 0.001, seed = 6150533)
mod <- sparseVAR(Y=scale(dat$Y))
f <- fitted(mod)
}
