% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{fitted.bigtime.VARX}
\alias{fitted.bigtime.VARX}
\title{Gives the fitted values of a model estimated using sparseVARX}
\usage{
\method{fitted}{bigtime.VARX}(object, ...)
}
\arguments{
\item{object}{model estimated using sparseVARX}

\item{...}{Not currently used}
}
\value{
Returns a matrix of fitted values
}
\description{
Gives the fitted values of a model estimated using sparseVARX
}
