% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{fitted.bigtime.VARX}
\alias{fitted.bigtime.VARX}
\title{Gives the fitted values of a model estimated using
\code{\link{sparseVARX}}}
\usage{
\method{fitted}{bigtime.VARX}(object, ...)
}
\arguments{
\item{object}{Model estimated using \code{\link{sparseVARX}}}

\item{...}{Not currently used}
}
\value{
Returns a matrix of fitted values
data(varx.example)
varx <- sparseVARX(Y=scale(Y.varx), X=scale(X.varx), selection="cv")
fit <- fitted(varx)
}
\description{
Gives the fitted values of a model estimated using
\code{\link{sparseVARX}}
}
