% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursiveforecast.R
\name{is.stable}
\alias{is.stable}
\title{Checks whether a VAR is stable}
\usage{
is.stable(mod, verbose = FALSE)
}
\arguments{
\item{mod}{model estimated using bigtime::sparseVAR. Can only be a model
with one coefficient vector. Hence, the model must either be estimated using
cv=TRUE or by giving a single lambda value}

\item{verbose}{If TRUE, then the actual maximum eigenvalue of the companion
matrix will be printed to the console. Default is FALSE}
}
\value{
Returns TRUE if the VAR is stable and FALSE otherwise
}
\description{
Using a model estimated by bigtime::sparseVAR, this function checks whether
the resulting VAR is stable
}
