% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagmatrix.R
\name{lagmatrix}
\alias{lagmatrix}
\title{Creates lagmatrix of estimated coefficients}
\usage{
lagmatrix(fit, model, returnplot = F)
}
\arguments{
\item{fit}{Fitted var, varx or varma model.}

\item{model}{Type of model that was estimated: var, varx or varma.}

\item{returnplot}{True or False: return plot of Lhat or not.}
}
\value{
A list with the following components
\item{Lhat}{Estimated lag matrix in the VAR model, or lag matrices in the VARX or VARMA model. The rows contain the responses, the columns contain the predictors.}
}
\description{
Creates lagmatrix of estimated coefficients
}
\examples{
data(Y)
data(X)
varxfit <- sparsevarx(Y=Y, X=X) # sparse VARX
lhats <- lagmatrix(fit=varxfit, model="varx")
}
