% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cv.R
\name{plot_cv}
\alias{plot_cv}
\title{Plot the Cross Validation Error Curve for a Sparse VAR or VARX}
\usage{
plot_cv(fit, model, ...)
}
\arguments{
\item{fit}{Fitted VAR or VARX model.
returned by \code{\link{sparseVAR}} or \code{\link{sparseVARX}}.}

\item{model}{Type of model that was estimated: VAR or VARX.}
}
\description{
Plot the Cross Validation Error Curve for a Sparse VAR or VARX
}
