% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursiveforecast.R
\name{reduce_cube}
\alias{reduce_cube}
\title{Reduces the third dimension of a cube and returns a data frame}
\usage{
reduce_cube(cube, dim3_names, name = deparse(substitute(dim3_names)))
}
\arguments{
\item{cube}{some 3D array}

\item{dim3_names}{vector of length dim(cube)[3] containing a name for
each slice}

\item{name}{what should the column containing the dim3_names be called?}
}
\value{
Returns a data frame contructed from the cube
}
\description{
This is only meant to be a helper function and not meant to be called by
the user itself
}
