% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvarx.R
\name{sparsevarx}
\alias{sparsevarx}
\title{Sparse estimation of the Vector AutoRegressive  with exogenous variables X (VARX) model}
\usage{
sparsevarx(Y, X, p = NULL, s = NULL, VARXpen = "HLag",
  VARXlPhiseq = NULL, VARXPhigran = NULL, VARXlBseq = NULL,
  VARXBgran = NULL, VARXalpha = 0, h = 1, cvcut = 0.9, eps = 10^-3)
}
\arguments{
\item{Y}{A \eqn{T} by \eqn{k} matrix of time series. If k=1, a univariate autoregressive model is estimated.}

\item{X}{A \eqn{T} by \eqn{m} matrix of time series.}

\item{p}{User-specified maximum endogenous autoregressive lag order. Typical usage is to have the program compute its own maximum lag order based on the time series length.}

\item{s}{User-specified maximum exogenous autoregressive lag order. Typical usage is to have the program compute its own maximum lag order based on the time series length.}

\item{VARXpen}{"HLag" (hierarchical sparse penalty) or "L1" (standard lasso penalty) penalization in VARX.}

\item{VARXlPhiseq}{User-specified grid of values for regularization parameter corresponding to the endogenous autoregressive coefficients in the VARX. Typical usage is to have the program compute
its own grid. Supplying a grid of values overrides this. WARNING: use with care.}

\item{VARXPhigran}{User-specified vector of granularity specifications for the penalty parameter grid corresponding to the endogenous autoregressive coefficients in the VARX:  First element specifies
how deep the grid should be constructed. Second element specifies how many values the grid should contain.}

\item{VARXlBseq}{User-specified grid of values for regularization parameter corresponding to the exogenous autoregressive coefficients in the VARX. Typical usage is to have the program compute
its own grid. Supplying a grid of values overrides this. WARNING: use with care.}

\item{VARXBgran}{User-specified vector of granularity specifications for the penalty parameter grid corresponding to the exogenous autoregressive coefficients in the VARX:  First element specifies
how deep the grid should be constructed. Second element specifies how many values the grid should contain.}

\item{VARXalpha}{a small positive regularization parameter value corresponding to squared Frobenius penalty. The default is zero.}

\item{h}{Desired forecast horizon in time-series cross-validation procedure.}

\item{cvcut}{Proportion of observations used for model estimation in the time series cross-validation procedure. The remainder is used for forecast evaluation.}

\item{eps}{a small positive numeric value giving the tolerance for convergence in the proximal gradient algorithm.}
}
\value{
A list with the following components
\item{Y}{\eqn{T} by \eqn{k} matrix of endogenous time series.}
\item{X}{\eqn{T} by \eqn{m} matrix of exogenous time series.}
\item{k}{Number of endogenous time series.}
\item{m}{Number of exogenous time series.}
\item{p}{Maximum endogenous autoregressive lag order of the VARX.}
\item{s}{Maximum exogenouss autoregressive lag order of the VARX.}
\item{Phihat}{Matrix of estimated endogenous autoregressive coefficients.}
\item{Bhat}{Matrix of estimated exogenous autoregressive coefficients.}
\item{phi0hat}{vector of VARX intercepts.}
}
\description{
Sparse estimation of the Vector AutoRegressive  with exogenous variables X (VARX) model
}
\examples{
data(Y)
data(X)
varxfit <- sparsevarx(Y=Y, X=X) # sparse VARX
Y1 <- matrix(Y[,1], ncol=1)
arxfit <- sparsevarx(Y=Y1, X=X) # sparse ARX
}
\references{
Wilms Ines, Sumanta Basu, Bien Jacob and Matteson David S. (2017), "Interpretable Vector AutoRegressions with
Exogenous Time Series" arXiv preprint.
}
\seealso{
\link{lagmatrix} and \link{directforecast}
}
